Move conversion functions to a dedicated file
